+++ dll/win32/ucrtbase/CMakeLists.txt
@@ -21,8 +21,12 @@ set_entrypoint(ucrtbase __acrt_DllMain 12)
 target_link_libraries(ucrtbase
     ucrt
     ucrtsupport
+    vcruntime
     wine
 )
+
+# Implicitly link to vcstartup
+target_link_libraries(libucrtbase vcstartup)
 
 if(MSVC)
     target_link_libraries(ucrtbase runtmchk)
