+++ sdk/lib/CMakeLists.txt
@@ -55,6 +55,7 @@ add_subdirectory(tzlib)
 add_subdirectory(ucrt)
 add_subdirectory(udmihelp)
 add_subdirectory(uuid)
+add_subdirectory(vcruntime)
 add_subdirectory(wdmguid)
 
 else()
