+++ sdk/lib/ucrt/CMakeLists.txt
@@ -102,7 +102,6 @@ include(stdio/stdio.cmake)
 include(stdlib/stdlib.cmake)
 include(string/string.cmake)
 include(time/time.cmake)
-include(vcruntime/vcruntime.cmake)
 
 add_library(ucrt OBJECT
     ${UCRT_CONIO_SOURCES}
@@ -128,5 +127,5 @@ add_library(ucrt OBJECT
     ${UCRT_VCRUNTIME_SOURCES}
 )
 
-#target_link_libraries(ucrt pseh)
+target_link_libraries(ucrt vcruntime)
 add_dependencies(ucrt psdk asm)
