+++ sdk/lib/vcruntime/__report_gsfailure.c
