+++ sdk/lib/vcruntime/__vcrt_init.c
