+++ sdk/lib/vcruntime/amd64/__security_check_cookie.s
