+++ sdk/lib/vcruntime/i386/__security_check_cookie.s
