+++ sdk/lib/vcruntime/inc/internal_shared.h
