+++ sdk/lib/vcruntime/inc/nt.h
