+++ sdk/lib/vcruntime/inc/ntrtl.h
