+++ sdk/lib/vcruntime/inc/nturtl.h
