+++ sdk/lib/vcruntime/initializers.cpp
